Fix type checking of let.
See <https://github.com/sdiehl/write-you-a-haskell/pull/89>. The problem
could be manifested by trying to type check:

    (let ((f (λ f (+ f 1)))) (f "foo"))

which would pass type checking and fail at runtime. The cause was that
when we did

    t1 <- infer e1

we would get a type variable along with some constraints on it. Like
'TVar e where e ~ Float'. But then we generalized e to `forall e. e`
without knowing it unified with Float, and that's totally different.