Partial implementation: check pattern type signatures.
This only seems to work if `inT` is sufficiently constrained in the
beginning. So,

    (if~ 3 (: $x $a) x ...) # fails as expected
    (if~ n (: $x $a) (+ x 1) ...) # should fail, doesn't
    (+ 1 (if~ n (: $x $a) x ...)) # should fail, doesn't

GHC gets both of these cases.

I think the problem is that I haven't implemented constraints.

It might work to do `inferTypedOn (const inT) (inferPat inT)  pat`
instead, and not return the second result from `inferPat`. Dunno if it
would be better. If this whole thing needs rethinking anyway, doesn't
matter much.